functions/ps2bq/requirements.txt
google-cloud-error-reporting==0.30.1
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1
pytz==2018.7